pyproject.toml
[tool.poetry]
name = "intershipback"
version = "0.1.0"
description = ""
authors = ["Mizhik <vlad.mezhinskiy07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
redis = "^5.0.7"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
celery = "^5.4.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."